remove s3 secret reconciliation unittests
