+++ pkg/3scale/amp/operator/highavailability_test.go
@@ -6,9 +6,25 @@ import (
 	"testing"
 
 	"github.com/3scale/3scale-operator/pkg/3scale/amp/component"
+	v1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 )
+
+func getTestSecret(namespace, secretName string, data map[string]string) *v1.Secret {
+	secret := &v1.Secret{
+		TypeMeta: metav1.TypeMeta{APIVersion: "v1", Kind: "Secret"},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      secretName,
+			Namespace: namespace,
+		},
+		StringData: data,
+		Type:       v1.SecretTypeOpaque,
+	}
+	secret.Data = getSecretDataFromStringData(secret.StringData)
+	return secret
+}
 
 func backendRedisTestData() map[string]string {
 	return map[string]string{
