+++ pkg/3scale/amp/operator/types.go
@@ -67,12 +67,6 @@ type OperatorApicastOptionsProvider struct {
 	Client         k8sclient.Client
 }
 
-type OperatorS3OptionsProvider struct {
-	APIManagerSpec *appsv1alpha1.APIManagerSpec
-	Namespace      string
-	Client         k8sclient.Client
-}
-
 type OperatorHighAvailabilityOptionsProvider struct {
 	APIManagerSpec *appsv1alpha1.APIManagerSpec
 	Namespace      string
